backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose
redis
celery
python-magic
pyyaml
requests
reportlab
docker
aiofiles
slowapi